Added optional preprocessing arguments
Illumina sequencing read length was restricting ability to identify products at 13 bases or longer. The result was that many fix2 and fix3 sequences were truncated in the Illumina reads. To adjust for this, this updated ProcessSamples allows additional columns to be added to the input excel file, e.g. columns "fix2" and "fix3" to define shorter fix2 and fix3 sequences. In this way, longer products can be identified up to 18 bases long.